requirements.txt
aiofiles==23.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
html5tagger==1.3.0
httptools==0.6.1
idna==3.6
multidict==6.0.5
pymongo==4.6.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sanic==23.12.1
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==69.1.1
tracerite==1.1.1
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvloop==0.19.0
websockets==12.0
